Compact gate works and integration tests pass
