+++ ipa-core/src/protocol/basics/sum_of_product/mod.rs
@@ -4,15 +4,16 @@ use crate::{
     error::Error,
     ff::Field,
     protocol::{
-        context::{Context, UpgradedMaliciousContext},
+        context::{Context},
         RecordId,
     },
     secret_sharing::replicated::{
-        malicious::{AdditiveShare as MaliciousReplicated, ExtendableField},
+        malicious::{ExtendableField},
         semi_honest::AdditiveShare as Replicated,
     },
 };
 
+#[cfg(feature = "descriptive-gate")]
 pub(crate) mod malicious;
 mod semi_honest;
 
@@ -43,19 +44,3 @@ impl<C: Context, F: Field> SumOfProducts<C> for Replicated<F> {
     }
 }
 
-#[async_trait]
-impl<'a, F: ExtendableField> SumOfProducts<UpgradedMaliciousContext<'a, F>>
-    for MaliciousReplicated<F>
-{
-    async fn sum_of_products<'fut>(
-        ctx: UpgradedMaliciousContext<'a, F>,
-        record_id: RecordId,
-        a: &[Self],
-        b: &[Self],
-    ) -> Result<Self, Error>
-    where
-        'a: 'fut,
-    {
-        malicious::sum_of_products(ctx, record_id, a, b).await
-    }
-}
