+++ ipa-core/src/protocol/boolean/bitwise_less_than_prime.rs
@@ -191,7 +191,7 @@ impl BitwiseLessThanPrime {
     }
 }
 
-#[derive(Step)]
+#[cfg_attr(feature = "descriptive-gate", derive(Step))]
 pub(crate) enum Step {
     CheckTrimmed,
     CheckIfAnyOnes,
