+++ ipa-core/src/protocol/context/upgrade.rs
@@ -162,12 +162,6 @@ where
     }
 }
 
-#[derive(Step)]
-pub(crate) enum Upgrade2DVectors {
-    #[dynamic(64)]
-    Upgrade2d(usize),
-}
-
 #[async_trait]
 impl<'a, C, F, I, M> UpgradeToMalicious<'a, I, Vec<M>> for UpgradeContext<'a, C, F, NoRecord>
 where
@@ -193,30 +187,6 @@ where
     }
 }
 
-#[async_trait]
-impl<'a, C, F, T, M> UpgradeToMalicious<'a, BitDecomposed<T>, BitDecomposed<M>>
-    for UpgradeContext<'a, C, F, RecordId>
-where
-    C: UpgradedContext<F>,
-    F: ExtendableField,
-    T: Send + 'static,
-    M: Send + 'static,
-    for<'u> UpgradeContext<'u, C, F, RecordId>: UpgradeToMalicious<'u, T, M>,
-{
-    async fn upgrade(self, input: BitDecomposed<T>) -> Result<BitDecomposed<M>, Error> {
-        let ctx_ref = &self.ctx;
-        let record_id = self.record_binding;
-        BitDecomposed::try_from(
-            self.ctx
-                .parallel_join(input.into_iter().enumerate().map(|(i, share)| async move {
-                    UpgradeContext::new(ctx_ref.narrow(&Upgrade2DVectors::Upgrade2d(i)), record_id)
-                        .upgrade(share)
-                        .await
-                }))
-                .await?,
-        )
-    }
-}
 
 #[async_trait]
 impl<'a, C, F> UpgradeToMalicious<'a, Replicated<F>, C::Share>
@@ -246,58 +216,6 @@ impl<F: Field, T: LinearSecretSharing<F>> IPAModulusConvertedInputRowWrapper<F, 
             trigger_value,
             _marker: PhantomData,
         }
-    }
-}
-
-#[derive(Step)]
-pub(crate) enum UpgradeModConvStep {
-    UpgradeModConv1,
-    UpgradeModConv2,
-    UpgradeModConv3,
-}
-
-#[async_trait]
-impl<'a, C, F>
-    UpgradeToMalicious<
-        'a,
-        IPAModulusConvertedInputRowWrapper<F, Replicated<F>>,
-        IPAModulusConvertedInputRowWrapper<F, C::Share>,
-    > for UpgradeContext<'a, C, F, RecordId>
-where
-    C: UpgradedContext<F>,
-    C::Share: LinearSecretSharing<F>,
-    F: ExtendableField,
-{
-    async fn upgrade(
-        self,
-        input: IPAModulusConvertedInputRowWrapper<F, Replicated<F>>,
-    ) -> Result<IPAModulusConvertedInputRowWrapper<F, C::Share>, Error> {
-        let (is_trigger_bit, trigger_value, timestamp) = try_join3(
-            self.ctx
-                .narrow(&UpgradeModConvStep::UpgradeModConv1)
-                .upgrade_one(
-                    self.record_binding,
-                    input.is_trigger_bit,
-                    ZeroPositions::Pvvv,
-                ),
-            self.ctx
-                .narrow(&UpgradeModConvStep::UpgradeModConv2)
-                .upgrade_one(
-                    self.record_binding,
-                    input.trigger_value,
-                    ZeroPositions::Pvvv,
-                ),
-            self.ctx
-                .narrow(&UpgradeModConvStep::UpgradeModConv3)
-                .upgrade_one(self.record_binding, input.timestamp, ZeroPositions::Pvvv),
-        )
-        .await?;
-
-        Ok(IPAModulusConvertedInputRowWrapper::new(
-            timestamp,
-            is_trigger_bit,
-            trigger_value,
-        ))
     }
 }
