+++ ipa-core/src/protocol/context/validator.rs
@@ -13,10 +13,10 @@ use crate::{
     ff::Field,
     helpers::Direction,
     protocol::{
-        basics::{check_zero, Reveal},
+        basics::{Reveal},
         context::{
-            Base, Context, MaliciousContext, SemiHonestContext, UpgradableContext,
-            UpgradedMaliciousContext, UpgradedSemiHonestContext,
+            Base, Context, SemiHonestContext, UpgradableContext,
+            UpgradedSemiHonestContext,
         },
         prss::SharedRandomness,
         RecordId,
@@ -28,6 +28,9 @@ use crate::{
     },
     sync::{Arc, Mutex, Weak},
 };
+
+#[cfg(feature = "descriptive-gate")]
+use crate::protocol::context::{MaliciousContext, UpgradedMaliciousContext};
 
 #[async_trait]
 pub trait Validator<B: UpgradableContext, F: ExtendableField> {
@@ -69,7 +72,7 @@ impl<F: ExtendableField> Debug for SemiHonest<'_, F> {
 
 /// Steps used by the validation component of malicious protocol execution.
 /// In addition to these, an implicit step is used to initialize the value of `r`.
-#[derive(Step)]
+#[cfg_attr(feature = "descriptive-gate", derive(Step))]
 pub(crate) enum Step {
     /// For the execution of the malicious protocol.
     MaliciousProtocol,
@@ -77,7 +80,7 @@ pub(crate) enum Step {
     Validate,
 }
 
-#[derive(Step)]
+#[cfg_attr(feature = "descriptive-gate", derive(Step))]
 pub(crate) enum ValidateStep {
     /// Propagate the accumulated values of `u` and `w`.
     PropagateUAndW,
@@ -194,6 +197,7 @@ impl<F: ExtendableField> MaliciousAccumulator<F> {
     }
 }
 
+#[cfg(feature = "descriptive-gate")]
 pub struct Malicious<'a, F: ExtendableField> {
     r_share: Replicated<F::ExtendedField>,
     u_and_w: Arc<Mutex<AccumulatorState<F::ExtendedField>>>,
@@ -201,6 +205,7 @@ pub struct Malicious<'a, F: ExtendableField> {
     validate_ctx: Base<'a>,
 }
 
+#[cfg(feature = "descriptive-gate")]
 #[async_trait]
 impl<'a, F: ExtendableField> Validator<MaliciousContext<'a>, F> for Malicious<'a, F> {
     /// Get a copy of the context that can be used for malicious protocol execution.
@@ -233,7 +238,7 @@ impl<'a, F: ExtendableField> Validator<MaliciousContext<'a>, F> for Malicious<'a
             .validate_ctx
             .narrow(&ValidateStep::CheckZero)
             .set_total_records(1);
-        let is_valid = check_zero(check_zero_ctx, RecordId::FIRST, &t).await?;
+        let is_valid = crate::protocol::basics::check_zero(check_zero_ctx, RecordId::FIRST, &t).await?;
 
         if is_valid {
             // Yes, we're allowed to downgrade here.
@@ -245,6 +250,7 @@ impl<'a, F: ExtendableField> Validator<MaliciousContext<'a>, F> for Malicious<'a
     }
 }
 
+#[cfg(feature = "descriptive-gate")]
 impl<'a, F: ExtendableField> Malicious<'a, F> {
     #[must_use]
     #[allow(clippy::needless_pass_by_value)]
@@ -304,6 +310,7 @@ impl<'a, F: ExtendableField> Malicious<'a, F> {
     }
 }
 
+#[cfg(feature = "descriptive-gate")]
 impl<F: ExtendableField> Debug for Malicious<'_, F> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "MaliciousValidator<{:?}>", type_name::<F>())
