+++ ipa-core/src/protocol/ipa_prf/boolean_ops/share_conversion_aby.rs
@@ -18,6 +18,7 @@ use crate::{
         SharedValue,
     },
 };
+use crate::protocol::basics::Reveal;
 
 #[derive(Step)]
 pub(crate) enum Step {
