+++ ipa-core/src/protocol/step/mod.rs
@@ -74,11 +74,3 @@ impl From<usize> for BitOpStep {
         Self::Bit(v)
     }
 }
-
-/// Set of steps that define the IPA protocol.
-#[derive(Step)]
-pub(crate) enum IpaProtocolStep {
-    /// Sort shares by the match key
-    #[dynamic(64)]
-    Sort(usize),
-}
